panago-price-validator/requirements.txt
playwright>=1.40.0
pandas>=2.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
pyyaml>=6.0
tenacity>=8.0
structlog>=23.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
